(Call_Editor): Instead of passing the template text into Call_Editor, pass a boolean indicating whether it should retrieve the template from the ComposeText widget and save it into the editor file.  This allows callers to tell Call_Editor to resave a modified template when in the middle of editing.  This is an enhancement, not a bug fix, and is necessitated by some other fixes.  Also update all invocations of Call_Editor for the new syntax.
If the Subject field is missing, add an empty Subject field to the
message and tell the user to fill it in and resend.  Also, if calling
Call_Editor to re-edit the message, tell it to resave the message into
the editor file.

If the Subject field is empty, tell the user to fill it in and resend.

If there is no Newsgroups field in the message, don't just put a
default Newsgroups field into the message and send it.  A program
should never post messages to newsgroups not explicitly requested by
the user.  Instead, add a default (or empty) Newsgroups field to the
message, and tell the user to edit it and then resend.

In composePane, free the header template as soon as it's no longer
needed -- a little earlier than it was being freed before, since it's
no longer being passed into Call_Editor.