getArticle() and moveToArticle() no longer display an error message when they can't get an article, so the caller has to do it.
Fix artListOldFunction() so that when multiple articles in a row
aren't available, it displays one message for all of them instead of
displaying a whole slew of messages.  This is important when the
active file on the news server isn't renumbered properly.  Problem
reported by Gerry.Tomlinson@newcastle.ac.uk.