Code cleanup: Don't define XRNrcsid if GCC_WALL is defined.
In Motif, use a RowColumn widget instead of a Form widget to put the
buttons in.

Code cleanup: Got rid of unused variables.

Moved the code for marking a button as the default into xmisc.c, so
that it can be used from the Information window code too.

Reenable an XtSetKeyboardFocus call for the Xaw interface; a comment
claimed that it was causing coredumps, but I doubt that's true
anymore, and in any case, it didn't say what X11 release it was
causing coredumps under, and the only way for me to find that out is
to reenable it.