Use $Date$ in the date header of the man page, so that the current RCS date shows up in the formatted man page.
Set up some variables to be updated automatically when the program is
built, and some files not to be included at all when they're not
actually used in the local installation.  See also the changes to
Imakefile to build xrn.man from xrn-man.src (to which this file has
been renamed).

Document unreadIconName and unreadIconPixmap.

Update the documentation for watchUnread to indicate what the list
separators are and to indicate that regular expressions, rather than
newsgroup names, should be listed.

Add documentation for the domainName resource.

Fix some typos, and some other minor formatting nits.