(fedUpART): Change the behavior of the "Fed up" button, so that it finds the next unread newsgroup even when "List old" has been selected in newsgroup mode.  Previously, if "List old" was selected in newsgroup mode, then hitting "Fed up" in article mode would exit article mode instead of going to vthe next unread newsgroup.
(artNextGroupFunction): Do the right thing even when "List old" has
been selected in newsgroup mode.  Previously, when the user clicked on
"Next group" and "List old" was selected in newsgroup mode, this
function would try to enter a whole bunch of newsgroups with no unread
articles; now, it checks for newsgroups with unread articles and only
tries to enters them.