Get rid of the code which prevents prefetching the next group when entering a newsgroup using "goto".  The prefetching is efficient enough that it doesn't hurt to prefetch the next group even if we're not going to go to it, and besides, if the user clicks on "Next group" after entering a group with "goto", he'll want the next group prefetched.
Modify getNearbyNewsgroup so that it copies the newsgroup on the line
whose position it's returning into the group_name array passed into
it.

Add code in the prefetch routines to put messages in the top info line
when a prefetch starts, finishes or is aborted.

When we prefetch a group and discover that it has no unread articles
in it, release its resources and prefetch the next group.  This means
that prefetching, once it starts, will continue until either it finds
a group with unread articles or it encounters a group that is too
large to prefetch.

Redraw the newsgroup text widget when done prefetching a newsgroup,
because the prefetching may have changed the widget's contents.

In prefetchNextUnreadGroup, instead of fetching the next unread group
after CurrentGroup, fetch the next unread group after the group passed
into it, or the first group if the group passed into it is null.

Delete an unnecessary prototype for prefetchNextGroup.

Use cancelPrefetch in a few places instead of doing its work
explicitly.

Add a prefetchGroup function, which takes a newsgroup name and tries
to prefetch that group.