Export redrawNewsgroupTextWidget so that it can be called by the prefetch routines when they're done prefetching a group.
Don't move the cursor forward in redrawNewsgroupTextWidget when
CurrentIndexGroup is empty.

Set the insertion point in redrawNewsgroupTextWidget instead of
updateNewsgroupMode.

Return from updateNewsgroupMode if we're not in newsgroup mode.

Prefetch the current newsgroup after updating newsgroup mode.  This
means that the current newsgroup will be prefetched when the program
starts up, when the exits article mode back to newsgroup mode, and
when a rescan is done.

When jumping to a newsgroup with "goto", copy the newsgroup into
CurrentIndexGroup, so that when the user switches back into newsgroup
mode, his position in the newsgroup list will change.  This makes the
behavior similar to that of rn.