Break up the prefetching work further, so that single chunks of prefetching work don't take more than a few seconds.  Dynamically ajust how much work to do in any single chunk based on how quickly the NNTP server is responding to requests.  Aim for a delay time of one second for any one prefetch work chunk.  In other words, when you're reading a newsgroup and XRN is prefetching the next one, you should never have to wait more than a few seconds in response to a command you give the program.
Change some "infoNow" calls in prefetch work procedures into "info"
calls, because they were causing deadlock.

Get rid of a minor memory leak if a prefetch is cancelled in between
when the prefetchSetFetch work procedure is installed and when it
actually runs.

Allow prefetchMax to be 0, in which case there is no limit on
newsgroup size when prefetching.

Mkae the SetupStage variable a global static instead of a static in
prefetchSetupUnreadGroup, and initialize it when starting a prefetch,
to avoid problems with cancelling a prefetch in the middle of
prefetchSetupUnreadGroup's work.