In info(), cache the last string displayed, to avoid unnecessary interactions with the X server.  Also, use XtVaSetValues() instead of XtSetValues().
In infoNow(), use xthHandlePendingExposeEvents() instead of
xthHandleAllPendingEvents(), so that we don't handle things like
button presses that will confuse us.  I had to do this because the
prefetching sequence was getting corrupted when it used infoNow() to
display its current status, and the user had hit a button that caused
a new prefetch to be started.