In topInfoHandler() and setTopInfoLine(), use the info() function to set the top info line, instead of doing it ourselves.  This is important because info() now caches what's in the info line.
In swapMode(), free NewsGroupsString and set it to 0 when leaving
newsgroup mode.  This is important because it enables
redrawNewsgroupTextWidget() to know when it really has to redraw the
widget and when it doesn't (to prevent unnecessary trips to the X
server).

In redrawNewsgroupTextWidget(), only redraw the text widget if if has
changed since the last time we drew it (or if we are just entering
newsgroup mode).

In artNextGroupFunction(), generate our own group list if
NewsGroupsString is null, and only free it if we generated it.

In artNextGroupFunction(), don't pass CurrentIndexGroup into
getNearbyNewsgroup(), because we're not actually using the real
NewsGroupsString, and the one we're using may not have the newsgroup
we're currently in in it, which will cause the first iteration through
the loop to skip a newsgroup that it shouldn't.