Modify finishPrefetch() so that instead of using timeouts and letting XtAppProcessEvent call the work procedures, we keep track of which work procedure we're using right now and call it directly.  This should speed things up and prevent some recursion problems.
In killArticles(), be more careful to set newsgroup->max_killed
properly on startup.

Move the messages about processing KILL files from killArticles() into
checkKillFiles().

In checkKillFiles(), do the right thing when resetting
newsgroup->max_killed after killing.

Rework a whole bunch of DEBUG messages, mostly in the prefetch
routines.

In all of the prefetch work procedures, return immediately at the
beginning of the procedure if the newsgroup currently being prefetched
is not the one we're expecting.  In order to make it possible for
thyem to determine that, pass the expected newsgroup into all of the
work procedures when setting them up.