Add a doPrefetch() action procedure, to be used in newsgroup mode to prefetch the current newsgroup.  (Already added calls to it in XRn.src and XRnMotif.src.)
Add an argument to updateSubjectWidget() to specify whether or not it
should prefetch when it's done updating.  Say yes when it's called
from functions that aren't the result of user button or key events,
and no from functions that are (because in that case, the prefetching
can be done from resources using the action procedure).

Call getSelection() with its new argument, which specifies whether or
not to clear the selection when done using it.

Remove the calls to prefetchGroup() in ngNextFunction and
ngPrevFunction(), because they're handled from action procedures now.