Merge info() and infoNow() into a single function called _info(), and change the prototypes for info() and infoNow() into macros that call _info().
Modify _info() so that it's reentrant and nested calls to it are ignored.
This way, if a call to infoNow() causes an event to be processed which causes
another call to _info(), the nested call to _info() doesn't blot out the
message put up by the first call to _info().