Remove the item about adding and documenting key bindings, because it has been done.
The DONT_FORCE_BUTTON_TRANSLATIONS and matchTranslations functionality
has been removed.  It's not relevant as long as we're not supporting
Motif, and when we do put Motif back, we can put it back in a more
correct way, if necessary.

Add an item about making attribution lines customizable.

Add an item about documenting the widget tree in the man page.