In translations and accelerators, use symbolic names for keys, instead of their hexidecimal values.  In particular, use "space" instead of 0x20, "Linefeed" instead of 0xff0a, and "Return" instead of 0xff0d.
In previous versions of XRN, all modes used the widgets to display
data in -- the top text window, the top info line, the top button box,
the bottom text window, the bottom info line, and the bottom button
box.  This structure was problematic because it took advantage of
available screen real estate poorly (why should part of the XRN window
be blank and unused in all modes except article mode?).  It also made
it difficult to maintain the code.  Therefore, the code has been
reorganized so that each mode has its own "Paned" widget which fills
the entire XRN window when the mode is active.  There are probably all
kinds of little user-visible aspects of this change, but the big ones
are: (1) Screen real estate is utilized more fully -- for example, in
newsgroup mode, there's no longer a bottom text window, info line or
button box, which means that many more newsgroups can be listed in the
same space.  (2) The "Bindings" resources and command-line options
have been removed.  Instead, users should use the "translations"
resources for the various modes in order to set bindings.  The correct
resource to use for customizing each mode's translations is given in
the documentation for that mode in the man page.  (3) Because each
mode uses separate widgets, it is now possible to customize the
appearance of modes separately -- for example, you can use different
fonts for the newsgroup or article lists in different modes.  (4)
Because the widget hierarchy has changed, the names of some widgets
have changed, and therefore some user X resources which refer to those
widgets might no longer work.  See the XRN source code or the
application-defaults file to find out about the new names of the
various widgets.

The key bindings for the various modes have been updated.  A few key
bindings were changed or removed, but for the most part, the primary
goal was to add new key bindings so that much more can be done without
reaching for the mouse.  Furthermore, all default key bindings are now
documented in the man page.

New functionality has been added to make it possible to move around in
the subject list in article mode without the displayed article
following the subject cursor.  For the most part, this change will be
transparent to people who don't want to use this functionality.
However, the one non-transparent aspect of it is that the up and down
arrows no longer display the next and previous article; instead, they
move the cursor without changing the displayed article.  The 'P' and
'N' keys retain the behavior which used to be bound to the arrow keys.
Users who prefer the old behavior can restore it by putting the
following in their X resources:

	xrn*artFrame.subjects.translations: #overrid	\n\
		~Meta<Key>Up:	artPrev()		\n\
		~Meta<Key>Down:	artNext()		\n

For more information about this new functionality, see the
documentation about it near the beginning of the "ARTICLE MODE"
section of the man page.

Two new buttons, "allNext" and "allPrev", have been added in all mode
to move the cursor to the next or previous newsgroup.

Some resources that were hard-coded in mesg.c have been moved into
XRn.src.