In previous versions of XRN, all modes used the widgets to display data in -- the top text window, the top info line, the top button box, the bottom text window, the bottom info line, and the bottom button box.  This structure was problematic because it took advantage of available screen real estate poorly (why should part of the XRN window be blank and unused in all modes except article mode?).  It also made it difficult to maintain the code.  Therefore, the code has been reorganized so that each mode has its own "Paned" widget which fills the entire XRN window when the mode is active.  There are probably all kinds of little user-visible aspects of this change, but the big ones are: (1) Screen real estate is utilized more fully -- for example, in newsgroup mode, there's no longer a bottom text window, info line or button box, which means that many more newsgroups can be listed in the same space.  (2) The "Bindings" resources and command-line options have been removed.  Instead, users should use the "translations" resources for the various modes in order to set bindings.  The correct resource to use for customizing each mode's translations is given in the documentation for that mode in the man page.  (3) Because each mode uses separate widgets, it is now possible to customize the appearance of modes separately -- for example, you can use different fonts for the newsgroup or article lists in different modes.  (4) Because the widget hierarchy has changed, the names of some widgets have changed, and therefore some user X resources which refer to those widgets might no longer work.  See the XRN source code or the application-defaults file to find out about the new names of the various widgets.
The various lists of buttons have been reordered to be in the same
order as the man page's list.

Some "gcc -Wall" nits have been eliminated by initializing variables
that don't really need to be initialized, but that make gcc complain
about them possibly being uninitialized.

Some button/action functions have been updated so that they will
behave correctly when they are called when the subject cursor is on an
article other than the currently displayed one.

The code which turns the mouse pointer into a watch when XRN is busy
has been rewritten to use Xt commands to set the cursor, rather than
using Xlib commands directly.

The "Callbacks" variable associated with each button has been
eliminated, because it's no longer necessary given the reorganization
that was done in order to allow separate panes for each mode.

A new "InfoLine" interface for creating a one-line information label
(usually to be placed inside a pane) has been added and is used to
create information lines for the various modes.