Some resources that were hard-coded in mesg.c have been moved into XRn.src.
A new "InfoLine" interface for creating a one-line information label
(usually to be placed inside a pane) has been added and is used to
create information lines for the various modes.

The generic button code has been revamped, so that there is a generic
"ButtonBox" interface for creating a widget to contain a box of
buttons and for adding buttons to it.