Don't use sizeof to determine the length of a MSG string, since sizeof will return the size of the string pointer, not the length of the string.
Add an argument to ConfirmationBox to specify that the first argument,
rather than the second, should be considered the "CONTINUE" option.

Fix the ConfirmationBox calls that ask if the newsgroup list should be
retrieved, so that they tell ConfirmationBox that the first, rather
than the second, argument should be considered the "CONTINUE" option.