Move the code for parsing kill file entries into a separate function. In the process, make the following user-visible changes in functionality:
* Entries which contain a command other than 'j', 'm' or 's', are
no longer actually checked against articles.  That is, previously, XRN
would compare the regular expression of such an entry against all
unkilled articles, even though it didn't know what to do if the
expression matched.  Now, it'll simply report an error about an
unknown command and move on to the next entry.

* XRN now notices the "h" option before the colon in kill file
entries.  Its treatment of the option is thus far somewhat limited --
its only use now is to restrict XRN to searching either the Subject or
From line of a message rather than both.  See the man page for more
information.  The functionality of this option will (I hope) be
increased soon.