Support sorting of articles by date as well as by subject.
Support killing of articles based on "Date" fields in addition to
"Subject", "From" and "Newsgroups".

Define constants for the various prefetch stages, and use them instead
of hard-coding stage numbers all over the place.

Add arguments to setUpGroupIncremental() to specify the first and last
article numbers to fetch; if they are zero, they default to
newsgroup->current and newsgroup->last as before.

Support prefetching of Date fields.

Move the digits() function into utils.c and rename it utDigits().

Make subjectIndexLine() available to other files (i.e., undo its
"static" and declare it in internals.h).

Delete getSortedSubjects() and getUnSortedSubjects(), and modify
getSubjects() to use the art_sort_*() routines instead of using the
deleted functions.

Rename strip() to subjectStrip(), and make it available to other files
(i.e., undo its "static" and declare it in internals.h).

Update getPrevSubject() so that it uses fillUpArray() instead of
calling the get*list() functions explicitly.

Update fillUpArray() so that it takes a last article number to fetch;
if 0, it defaults to newsgroup->last as before.

Update fillUpArray() to use setUpGroupIncremental() instead of calling
get*list() functions directly.