Add newsgroup arguments to fillUpArray() and moveToArticle().
Update articleIsAvailable() so that it uses fillUpArray() instead of
gesubjectlist().  This is important to ensure that all the headers
besides Subject are fetched too (since the code assumes in numerous
places that if the subject is non-null, all the headers have been
fetched).

Rename the FETCH_STAGE symbols to HEADERS_STAGE instead.

Isolate the code for fetching headers in its own function, and use
that from fillUpArray() instead of calling setUpGroupIncremental(), to
avoid problems with the prefetched group getting reset when we don't
want it to be.