Move the findArticle() function, which is used only in article mode, from cursor.c into artMode.c.
Update findArticle() so that it always starts searching at position 0
(since that's how it was being used before anyway) and so that it has
the option of returning an error instead of exiting the program if it
can't find the requested article.

In the "Goto article" command, don't regenerate the Subject listing
unless the requested article isn't in the current list of subjects.