(get_data_from_server, etc.): Limit the longest line that can be read by get_data_from_server.  If we don't do this, then if the INN server screws up and starts sending us garbage with really long lines, or if someone posts an article with really long lines in the body, we'll hog memory trying to read the long lines in one chunk.
The line limit right now is 8Kb.  Article header lines longer than
that will be truncated when they are read from the server.  Article
body lines longer than that should be handled properly, i.e., without
any truncation or extra newlines inserted into them.

(get_data_from_server): Now handles properly the case when we read
everything except the final '\n' from the server in one call to
get_server, and then read the final '\n' in the next call after that.

(getarticle): The code for stripping backspace characters has been
rewritten to be clearer and more efficient.

(getarticle): When displayLocalTime is True, only "Date:" fields which
actually occur in the header of the message will be converted
(previously, XRN would attempt to do date conversion on any line in
the message starting with "Date:", even in the message body).