Fix the following bug in the thread-killing code: When "artKillThread" was executed on the first article in a thread, a correct entry for the thread would be placed into the KILL file, but none of the article's descendants would be killed correctly in the current session.
When the user selects "List old" display previously read articles
whose numbers are higher than the current first listed article.

(artResortFunction): Allow the type of sorting to be used to be
specified as an argument or arguments of this action procedure, so
that the user can create a key binding to temporarily change article
sorting.

Make the "Resort" button inactive when the user has not specified any
sorting.