When posting a followup or reply, make sure that the text of the original article can be included, even when it is canceled or the user exits the newsgroup before hitting the "include article" button in the composition pane.
Redo the way retrieved articles are cached.  Instead of keeping all
articles in the current newsgroup cached until the user exits the
group, automatically keep a rotating cache of articles, and delete old
articles in the cache only as their slots are needed.  Furthermore,
allow the number of files in the cache and the total size of the cache
to be configurable with the cacheFilesMaxFiles and cacheFilesMaxSize
resources.  This should allow users to read large newsgroups without
filling up their /tmp, as well as speeding up the transitions from
newsgroup to newsgroup (or from article mode to newsgroup mode)
because XRN no longer has to pause to delete all the articles in a
group when exiting it.