Fixed some typos.
Updated the section about where to get XRN, to indicate that it's no
longer at ftp.cam.ov.com and that xrn-linux.tgz contains the last
production release, not the current beta release.

Deleted the paragraph about comp.sources.x, which is dead.

Updated the section about the Motif interface.