Add bug to to-do list: It appears that if there are two messenger windows open, +the background tasks run separately in each of them.
Add to to-do list: When checking pending messages, shouldn't read
entire message into +memory until we've confirmed that it is time to
send it.

Header field matching should be case-insensitive (because, e.g.,
Exchange stores "X-Send-Later-At" in drafts folder as
"x-send-later-at").

Fix minor bug in debugging code in the previous checkin to make
interval changes take effect immediately.

Make toolbar update properly when it is added to compose window.