Add some debug logging code.
Fix a bug in the code for adding the custom headers required by this
add-on so that they are added properly the first time. I think this
fixes the problem with the Send Later column not displaying properly
in the Drafts folder for a while after the add-on is installed!