Upversion to 3.3.6beta5.
Delete some unused variables.

Restructure how window loads and unloads are handled so we're not
trying to execute code that requires the DOM before the DOM has
loaded. This fixes issues with Send Later 3 causing icons from other
add-ons to not show up in the status bar in the compose window.