Upversion to 3.3.7beta1.
Don't try to deliver scheduled messages when Thunderbird is in offline
mode. This means that messages scheduled to be delivered while you are
in offline mode won't be moved from your Drafts folder into your
Outbox until you go back online. It also means that the "Date" header
of these messages will be set to the time when you go back online, NOT
the scheduled delivery time. You can revert to the old behavior, i.e.,
messages moved into Outbox and dated as of scheduled delivery time, by
using the advanced configuration editor to set the preference
extensions.sendlater3.send_while_offline to true. I am interested in
hearing from users whether you prefer this new behavior or the old
behavior, so that I can decide which behavior should be the default
and whether there are enough proponents of both behaviors that I
should add a checkbox to the preferences window rather than forcing
people to use the advanced configuration editor to change the
preference.