Previous Thunderbird 15 workaround was not adequate. Refactor code to make it clearer and more robust and to fully address the Thunderbird 15 bug.
Ensure that mailing lists are expended properly before scheduled
messages are saved in Drafts folder.

Add a hidden preference extensions.sendlater3.logging.filter which can
be used to specify a regular expression to indicate which messages
should be logged.

Add a a default setting for the preference
extensions.sendlater3.checktimepref.is_milliseconds so that it shows
up in the advanced configuration editor even when it has not been set
explicitly by the user.