Each Send Later 3 compose window toolbar item can now be positioned separately in the toolbar, giving the user complete control over which items appear in the toolbar and the order in which they appear.
Since the Send Later 3 shortcut buttons are now added or removed by
customizing the compose window toolbar directly, there are no longer
checkboxes in the Send Later 3 preferences controlling whether the
shortcut buttons appear in the toolbar.

You can now add a date/time picker to the toolbar in addition to or
instead of the text input box.

The Send Later button in the pop-up prompt window is now activated, as
it should be, when the window first comes up when re-schduling a
previously scheduled draft.