Casts between integers and pointers need to use same size types
Determine at compile time whether the size of a pointer is equal to
the size of int or long, and use whichever of those matches when
casting to or from pointers types.

If the size of a pointer is equal to something other than the size of
int or long on some platform, then utils.h and configure.ac will need
to be updated (and autoreconf will need to be run) to accommodate
that.